Automatically build and ship documentation
Build documentation in debian/gem2deb_generated_docs/html, ship it in
/usr/share/doc/$package/html.

* dh_make_ruby: Create debhelper files for dirs, links, doc-base, and docs
  to install documentation automatically built by dh_ruby. Add authors
  function wrapping metadata.authors.

* dh_ruby: call documentation clean and build functions from installer.
  Build documentation by calling rdoc.

* installer: clean and build documentation. Introduce new attribute
  docdir and add :docdir to destdir function.

* metadata: add authors attribute.

* gem2deb: add DOC_DIR attribute.

* Add unit and integration tests.

  + Add documentation sample package.

NB! This does not suggest libjs-jquery, but leaves it to the user to
install if they want fancy presentation. This is because libjs-jquery
recommends javascript-common which in turn suggests apache | lighttpd |
httpd, leading to a lot of things automatically installed if suggests
was configured to install automatically.